requirements.txt
tensorflow-gpu>=2.4.0
numpy==1.19.2
pandas==1.1.3
glob2>=0.7
scikit-learn>=0.23.2
re
pickle
os
gc